fixing simulation again

+++ Graduation_Project/Data/Models/MachineTypeMonitoringAttribute.cs
@@ -8,8 +8,8 @@ namespace Graduation_Project.Data.Models
         public int MachineTypeId { get; set; }
         public MachineType MachineType { get; set; } = null!;
         public ICollection<MonitorAttributeAlertRule> AlertRules { get; set; } = [];
-        public int UpperRange { get; set; }
-        public int LowerRange { get; set; }
+        public double UpperRange { get; set; }
+        public double LowerRange { get; set; }
 
 
     }
